+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Sentence Module
 
 - Add `segment_text` function for breaking text into sentences with character-level indices
+- Add `corpus` submodule with input classes for text and TEI/XML and factory
+  method to initialize appropriate input class based on file type
 
 ### Application
 
-- Add initial placeholder application
+- Add initial application that can build sentence corpora for text files
 
 ### Documentation